Adding of git request api and separation of git request to won component
